+++ apps/decodex-app/Sources/DecodexApp/AccountPanelSelectionMetrics.swift
@@ -1,0 +1,60 @@
+extension AccountPanelView {
+	var codexAuthLabel: String {
+		guard let auth = store.accountList?.codexAuth else {
+			return "No Codex auth"
+		}
+
+		if emailsHidden {
+			if let account = account(matching: auth.selector) {
+				return displayName(for: account)
+			}
+			let identity = auth.accountFingerprint.isEmpty ? auth.selector : auth.accountFingerprint
+			return AccountDisplay.alias(forIdentity: identity)
+		}
+
+		return AccountDisplay.compactEmail(auth.displayName)
+	}
+
+	var decodexModeLabel: String {
+		guard let control = store.accountList?.control else {
+			return "Not loaded"
+		}
+
+		if let selector = control.accountSelector, selector.isEmpty == false {
+			if emailsHidden {
+				if let account = account(matching: selector) {
+					return "To \(displayName(for: account))"
+				}
+
+				return "To \(AccountDisplay.alias(forIdentity: selector))"
+			}
+
+			if selector.contains("@") {
+				return "To \(AccountDisplay.compactEmail(selector))"
+			}
+
+			return "To \(AccountDisplay.compactIdentity(selector))"
+		}
+
+		return control.mode
+	}
+
+	var hasFixedSelection: Bool {
+		guard let selector = store.accountList?.control.accountSelector else {
+			return false
+		}
+
+		return selector.isEmpty == false
+	}
+
+	var headerSubtitle: String {
+		let count = store.accounts.count
+		let accountLabel = "\(count) account\(count == 1 ? "" : "s")"
+		let routeLabel = hasFixedSelection ? "Routed" : "Balanced"
+		return "\(accountLabel) · \(routeLabel)"
+	}
+
+	var emailsHidden: Bool {
+		accountPrivacy != AccountPrivacy.visibleValue
+	}
+}
