+++ apps/decodex-app/Sources/DecodexApp/PanelWindowSizing.swift
@@ -1,23 +1,11 @@
 import AppKit
 import SwiftUI
 
-private struct PanelWindowContentSizeKey: PreferenceKey {
-	static let defaultValue = CGSize.zero
-
-	static func reduce(value: inout CGSize, nextValue: () -> CGSize) {
-		let next = nextValue()
-		guard next != .zero else {
-			return
-		}
-		value = next
-	}
-}
-
 private struct PanelWindowSizeReporter: NSViewRepresentable {
 	let contentSize: CGSize
 
 	func makeNSView(context: Context) -> NSView {
-		let view = PanelWindowSizingView(frame: .zero)
+		let view = PanelWindowSizingProbeView(frame: .zero)
 		view.didMoveToWindow = { [weak coordinator = context.coordinator] view in
 			coordinator?.retryResizeAfterWindowAttachment(from: view)
 		}
@@ -95,67 +83,6 @@ private struct PanelWindowSizeReporter: NSViewRepresentable {
 		private static func sizeDiffers(_ lhs: NSSize, _ rhs: NSSize) -> Bool {
 			abs(lhs.width - rhs.width) > 0.5 || abs(lhs.height - rhs.height) > 0.5
 		}
-	}
-}
-
-private final class PanelWindowSizingView: NSView {
-	var didMoveToWindow: ((PanelWindowSizingView) -> Void)?
-
-	override func viewDidMoveToWindow() {
-		super.viewDidMoveToWindow()
-		didMoveToWindow?(self)
-	}
-}
-
-enum PanelWindowSizingLayout {
-	private static let placementMargin: CGFloat = 8
-
-	static func roundedContentSize(for contentSize: CGSize) -> NSSize {
-		NSSize(
-			width: ceil(contentSize.width),
-			height: ceil(contentSize.height)
-		)
-	}
-
-	static func frame(
-		forContentSize contentSize: NSSize,
-		currentFrame: NSRect,
-		frameSizeForContentSize: (NSSize) -> NSSize,
-		visibleFrame: () -> NSRect?
-	) -> NSRect {
-		let frameSize = frameSizeForContentSize(contentSize)
-		let proposedFrame = NSRect(
-			x: currentFrame.midX - frameSize.width / 2,
-			y: currentFrame.maxY - frameSize.height,
-			width: frameSize.width,
-			height: frameSize.height
-		)
-		guard let visibleFrame = visibleFrame() else {
-			return proposedFrame
-		}
-
-		return NSRect(
-			x: clamped(
-				proposedFrame.origin.x,
-				min: visibleFrame.minX + placementMargin,
-				max: visibleFrame.maxX - frameSize.width - placementMargin
-			),
-			y: clamped(
-				proposedFrame.origin.y,
-				min: visibleFrame.minY + placementMargin,
-				max: visibleFrame.maxY - frameSize.height - placementMargin
-			),
-			width: frameSize.width,
-			height: frameSize.height
-		)
-	}
-
-	private static func clamped(_ value: CGFloat, min: CGFloat, max: CGFloat) -> CGFloat {
-		guard min <= max else {
-			return min
-		}
-
-		return Swift.min(Swift.max(value, min), max)
 	}
 }
